Testes-Aceite/requirements.txt
robotframework==5.0.1
robotframework-seleniumlibrary >= 3.3.1
robotframework-debuglibrary
